WNB-25: Move all sequelize-related stuff into db folder
