--- .env
@@ -1,2 +1,0 @@
-process.env.VAR1="Some value"
-process.env.VAR2="Another Value"
