+++ README.md
@@ -10,20 +10,16 @@ This is a repo for the API part of Wannabook.
 
 ### Server Setup
 * Run `npm i` to install dependencies
-* Create a `config.json` file inside `/config`. This is responsible for syncing your models with database
+* Create initial DB and run migrations (see 'Working with database' below).
 * Run `npm run start:dev` to start server with nodemon
-* Add environment variables required to run the application to .env file in project root. Don't forget to puut quotes around values.
-```
-process.env.VAR1="Some value"
-process.env.VAR2="Another Value"
-```
+* Add environment variables required to run the application to one of .env files under the `env` folder.
 * Go to ```http://localhost:5000/```
 
 ### Working with database
-1. To generate initial DB and (or) run outstanding migrations, create `config.json` based on `config-example.json`. Put quotes around every entry except `null`s.
+1. To generate initial DB and (or) run outstanding migrations, create `config.js` based on `config-example.js`.
 
-2. Run `sequelize db:create` and it will create the DB 'Wannabook' as specified in `config.json`.
+2. Install Run `env-cmd -f ./env/dev.env sequelize db:create` and it will create the DB 'Wannabook' as specified in `config.js`.
 
-3. Inside of migrations folder, you will find all migrations. To run them, do `sequelize db:migrate` and only new migrations will run. They will update your DB accordingly.
+3. Inside of migrations folder, you will find all migrations. To run them, do `env-cmd -f ./env/dev.env sequelize db:migrate` and only new migrations will run. They will update your DB accordingly.
 
-4. Fill the DB with test date by running `npx sequelize db:seed:all`
+4. Fill the DB with test date by running `env-cmd -f ./env/dev.env npx sequelize db:seed:all`
