+++ app.js
@@ -2,9 +2,7 @@ const express = require('express');
 const bodyParser = require('body-parser');
 
 const test = require('./routes/test');
-const config = require('./configuration/config');
-const models = require('./models');
-require('./.env');
+const models = require('./db/models');
 
 const app = express();
 
@@ -15,8 +13,7 @@ app.get('/', (req, res) => res.json({ msg: 'Wannabook!' }));
 
 app.use('/test', test);
 
-const devPort = config.app.port;
-const port = process.env.PORT || devPort;
+const port = process.env.PORT || 5000;
 
 models.sequelize.sync().then(function() {
   app.listen(port, () => console.log(`Server running on port ${port}`));
