--- configuration/config.js
@@ -1,7 +1,0 @@
-const config = {
-    app: {
-        port: 5000
-    }
-};
-
-module.exports = config;
