+++ db/config/config-example.js
@@ -1,0 +1,23 @@
+module.exports = {
+  development: {
+    username: 'root',
+    password: null,
+    database: 'wannabook',
+    host: '127.0.0.1',
+    dialect: 'mysql',
+  },
+  staging: {
+    username: process.env.DB_USERNAME,
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_NAME,
+    host: process.env.DB_HOSTNAME,
+    dialect: 'mysql',
+  },
+  production: {
+    username: process.env.DB_USERNAME,
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_NAME,
+    host: process.env.DB_HOSTNAME,
+    dialect: 'mysql',
+  },
+};
